Completed Build section of tutorial
Note to self: tried `npm test` as part of the setup but all the
tests were failing. Reading these suggest that they are testing
for elements on the page being there. These elements are part
of the Build section requirements and, what do you know, with this
commit, the tests are passing.

    --- Lionel Saliou, Ph.D